refactor: split types.ts into domain-scoped modules
Move types from the monolithic src/types.ts into focused
modules under src/types/: config, docker, policy, logging,
and pid. A barrel index.ts re-exports everything for full
backwards compatibility.

Closes #2475
